+++ web/src/pages/index.tsx
@@ -473,6 +473,7 @@ const Index: React.FC<IndexProps> = ({}) => {
               <Heading as="h2" size="md" mt={2}>
                 Mes documents
               </Heading>
+              
             </GridItem>
           </Grid>
         </Center>