+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - users can't access website while they're not accepted or if they are refused
 - setting convs to inactive
 - Uploading files using FilePond (admin uploads document for selected services & postes, file is stored in server filesystem. Path, name and services / postes are stored in postgres)
+- documents GET route
 
 ## [v0.0.4](https://github.com/benzinho75/JEECE-collab/compare/v0.0.3...v0.0.4) - 2021-03-12